% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_SWATmgt.R
\name{write.SWATmgt}
\alias{write.SWATmgt}
\title{write.SWATmgt}
\usage{
write.SWATmgt()
}
\description{
\code{write.SWATmgt} reads the management operations from the selected csv
file and writes them to the management files in the selected txtInOut folder.
}

